feat: resume module restructure
