+++ src/app/(admin)/resume/page.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import ResumeBuilderCard from "@resume-builder/components/resume-builder";
+import ResumeBuilderCard from "@resume/components/resume-builder";
 import AsyncWrapper from "@shared/components/extends/async-wrapper";
 import { useResumes } from "@shared/queries/resume";
