+++ src/modules/resume/components/section-education.tsx
@@ -1,6 +1,6 @@
-import { TowerControl } from "lucide-react";
+import { School } from "lucide-react";
 
-import SectionWrapper from "@resume-builder/components/section-wrapper";
+import SectionWrapper from "@resume/components/section-wrapper";
 import OpenAIPrompt from "@shared/components/extends/openai-prompt";
 import { Input } from "@shared/components/ui/input";
 import { Label } from "@shared/components/ui/label";
@@ -8,7 +8,7 @@ import { Textarea } from "@shared/components/ui/textarea";
 
 export default function SectionExperience() {
   return (
-    <SectionWrapper header="Work experience" icon={<TowerControl />}>
+    <SectionWrapper header="Education" icon={<School />}>
       <div className="flex flex-col justify-start items-stretch gap-4">
         <div className="grid w-full items-center gap-1.5">
           <Label htmlFor="fullName">Company</Label>
