+++ src/shared/repositories/resume.ts
@@ -3,25 +3,25 @@ import { Resume } from "@shared/types/resume"
 
 export const resumeRepository = {
   getAll: () => {
-    return httpClient('/resume')
+    return httpClient('/resumes')
   },
   getById: (id: string) => {
-    return httpClient(`/resume/${id}`)
+    return httpClient(`/resumes/${id}`)
   },
   create: (data: Resume) => {
-    return httpClient('/resume', {
+    return httpClient('/resumes', {
       method: 'POST',
       body: JSON.stringify(data),
     })
   },
   update: (id: string, data: Resume) => {
-    return httpClient(`/resume/${id}`, {
+    return httpClient(`/resumes/${id}`, {
       method: 'PUT',
       body: JSON.stringify(data),
     })
   },
   delete: (id: string) => {
-    return httpClient(`/resume/${id}`, {
+    return httpClient(`/resumes/${id}`, {
       method: 'DELETE',
     })
   }
